+++ src/test_main.rs
@@ -4,8 +4,8 @@ use siamesedb::DbList;
 use siamesedb::DbMap;
 
 fn main() {
-    _test_a1();
-    //_test_a2();
+    //_test_a1();
+    _test_a2();
 }
 fn _test_a1() {
     //_test00_map();
@@ -17,6 +17,7 @@ fn _test_a1() {
     //
     // 1m
     //
+    /*
     #[rustfmt::skip]
     _test_db_map(
         TestC { max_cnt: 1_000, f_get: false, f_delete: false, f_repeat: 1 },
@@ -24,7 +25,6 @@ fn _test_a1() {
         |i: usize| { format!("key{:01}", i) },
         |i: usize| { format!("value{:01}", i) },
     );
-    /*
     #[rustfmt::skip]
     _test_db_map(
         TestC { max_cnt: 1_000, f_get: true, f_delete: true, f_repeat: 3 },
@@ -95,11 +95,17 @@ fn _test_a2() {
     // 10k
     //
     // 0.35user 0.01system 0:00.38elapsed 93%CPU (0avgtext+0avgdata 2208maxresident)k
+    //
+    // finish put: 14.096µs/op
+    // db_map.depth_of_node_tree(): 5
+    // finish get: 3.341µs/op
+    // finish delete: 9.314µs/op
+    // 0.26user 0.00system 0:00.27elapsed 96%CPU (0avgtext+0avgdata 2396maxresident)k
     /*
     #[rustfmt::skip]
     _test_db_map(
         TestC { max_cnt: 10_000, f_get: true, f_delete: true, ..Default::default() },
-        CheckC { check: false, ..Default::default() },
+        CheckC { check: false, f_depth: true, ..Default::default() },
         |i: usize| { format!("key{:01}", i) },
         |i: usize| { format!("value{:01}", i) },
     );
@@ -108,33 +114,57 @@ fn _test_a2() {
     // 100k
     //
     // 4.57user 0.09system 0:04.78elapsed 97%CPU (0avgtext+0avgdata 2332maxresident)k
-    /*
+    //
+    // finish put: 16.62µs/op
+    // db_map.depth_of_node_tree(): 6
+    // finish get: 4.021µs/op
+    // finish delete: 11.396µs/op
+    // 3.11user 0.07system 0:03.23elapsed 98%CPU (0avgtext+0avgdata 2272maxresident)k
     #[rustfmt::skip]
     _test_db_map(
         TestC { max_cnt: 100_000, f_get: true, f_delete: true, ..Default::default() },
-        CheckC { check: false, ..Default::default() },
+        CheckC { check: false, f_depth: true, ..Default::default() },
+        |i: usize| { format!("key{:01}", i) },
+        |i: usize| { format!("value{:01}", i) },
     );
-    */
+    /*
+     */
     //
     // 1m
     //
     // 54.47user 0.99system 0:56.10elapsed 98%CPU (0avgtext+0avgdata 2264maxresident)k
+    //
+    // finish put: 19.782µs/op
+    // db_map.depth_of_node_tree(): 7
+    // finish get: 4.714µs/op
+    // finish delete: 12.943µs/op
+    // 36.39user 0.78system 0:37.60elapsed 98%CPU (0avgtext+0avgdata 2336maxresident)k
     /*
     #[rustfmt::skip]
     _test_db_map(
         TestC { max_cnt: 1_000_000, f_get: true, f_delete: true, ..Default::default() },
-        CheckC { check: false, ..Default::default() },
+        CheckC { check: false, f_depth: true, ..Default::default() },
+        |i: usize| { format!("key{:01}", i) },
+        |i: usize| { format!("value{:01}", i) },
     );
-    */
+     */
     //
     // 10m
     //
     // 627.91user 12.64system 10:51.87elapsed 98%CPU (0avgtext+0avgdata 2344maxresident)k
+    //
+    // finish put: 22.899µs/op
+    // db_map.depth_of_node_tree(): 8
+    // finish get: 4.981µs/op
+    // finish delete: 15.317µs/op
+    // 415.67user 11.13system 7:14.21elapsed 98%CPU (0avgtext+0avgdata 2444maxresident)k
     /*
     #[rustfmt::skip]
     _test_db_map(
         TestC { max_cnt: 10_000_000, f_get: true, f_delete: true, ..Default::default() },
-        CheckC { check: false, ..Default::default() },
+        CheckC { check: false, f_depth: true, ..Default::default() },
+        |i: usize| { format!("key{:01}", i) },
+        |i: usize| { format!("value{:01}", i) },
     );
     */
     //
@@ -142,11 +172,35 @@ fn _test_a2() {
     // 100m
     //
     // 7466.42user 229.81system 2:15:24elapsed 94%CPU (0avgtext+0avgdata 2224maxresident)k
+    //
+    // finish put: 27.208µs/op
+    // db_map.depth_of_node_tree(): 9
+    // finish get: 7.15µs/op
+    // finish delete: 22.514µs/op
+    // 4875.26user 231.29system 1:34:47elapsed 89%CPU (0avgtext+0avgdata 2448maxresident)k
+    //
+    // node cache size: 64
+    //
+    // record free: [(16, 0), (24, 0), (32, 0), (48, 0), (64, 0), (256, 0), (512, 0), (2048, 0)]
+    // record used: [(16, 100), (24, 999900), (32, 99000000), (48, 0), (64, 0), (256, 0), (512, 0), (2048, 0)]
+    // node free: [(32, 503), (72, 1), (104, 0), (144, 0), (176, 0), (216, 0), (232, 0), (256, 0)]
+    // node used: [(32, 0), (72, 8605028), (104, 5307029), (144, 2000655), (176, 0), (216, 0), (232, 0), (256, 0)]
+    // db_map.is_balanced(): true
+    // db_map.is_dense(): true
+    // record_size_stats(): [(16, 100), (24, 999900), (32, 99000000)]
+    //
+    // finish put: 27.041µs/op
+    // db_map.depth_of_node_tree(): 9
+    // finish get: 5.424µs/op
+    // finish delete: 17.437µs/op
+    // 5050.66user 153.20system 1:27:33elapsed 99%CPU (0avgtext+0avgdata 2556maxresident)k
     /*
     #[rustfmt::skip]
     _test_db_map(
         TestC { max_cnt: 100_000_000, f_get: true, f_delete: true, ..Default::default() },
-        CheckC { check: false, ..Default::default() },
+        CheckC { check: false, f_depth: true, ..Default::default() },
+        |i: usize| { format!("key{:01}", i) },
+        |i: usize| { format!("value{:01}", i) },
     );
     */
     // 3194.15user 44.34system 54:27.41elapsed 99%CPU (0avgtext+0avgdata 2408maxresident)k
@@ -420,6 +474,7 @@ struct TestC {
 #[derive(Debug, Default, Clone, Copy)]
 struct CheckC {
     check: bool,
+    f_depth: bool,
     f_mst: bool,
     f_graph: bool,
 }
@@ -452,6 +507,8 @@ where
         //
         if check_cnf.check {
             _print_check_db_map(&db_map, check_cnf);
+        } else if check_cnf.f_depth {
+            _print_depth_db_map(&db_map);
         }
         //
         if test_cnf.f_get {
@@ -603,5 +660,19 @@ fn _print_check_db_list(db_list: &FileDbList, check_cnf: CheckC) {
     println!(
         "record_size_stats(): {}",
         db_list.record_size_stats().unwrap()
+    );
+}
+
+fn _print_depth_db_map(db_map: &FileDbMap) {
+    println!(
+        "db_map.depth_of_node_tree(): {}",
+        db_map.depth_of_node_tree().unwrap()
+    );
+}
+
+fn _print_depth_db_list(db_list: &FileDbList) {
+    println!(
+        "db_list.depth_of_node_tree(): {}",
+        db_list.depth_of_node_tree().unwrap()
     );
 }